requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai>=0.8.0
python-docx==1.1.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
authlib
starlette
itsdangerous

# RAG System
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain>=0.1.0
tiktoken>=0.5.2
